crates/algo_models_ffi/tests/py/requirements.txt
cffi==1.17.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.4
